eval/requirements.txt
pandas
requests
python-dotenv
rouge-score
sqlparse
